attempting move the files back to the root
