--- custom_components/smartthinq/manifest.json
@@ -1,8 +1,0 @@
-{
-    "domain": "smartthinq",
-    "name": "hass-smartthinq",
-    "documentation": "https://github.com/sampsyo/hass-smartthinq",
-    "dependencies": [],
-    "codeowners": ["@sampsyo"],
-    "requirements": ["wideq>=1.0.1"]
-}
