+++ hacs.json
@@ -1,6 +1,6 @@
 {
   "name": "LG SmartThinQ Appliances",
-  "content_in_root": false,
+  "content_in_root": true,
   "domains": ["climate", "sensor"],
   "iot_class": "cloud_poll",
   "homeassistant": "0.96.0",
